Use specific placeholder value instead of true for bare values
Currently when asked to log "bare values" the emitter uses a value of `true` internally in the hashes it uses to build the log data, knowing that `true` values will be removed again when the data is serialized into a logfmt line. This change removes the overloading of `true` and instead uses a specific value to indicate a bare value. This allows the handling of genuine `true` values to be changed without affecting bare value behaviour.