Bring back `#disable!` support
As part of [#7](https://github.com/heroku/l2meter/pull/7) we lost the
ability to totally disable the logger, across all threads. It's unclear
if this was intentional, but I am guessing it was not. This will restore
the behavior by returning a new `NullOutput` instance when trying to
write while disabled.

This might cause A LOT of object allocations, all with short lives.
Hopefully. If so, we could consider some form of memoization of the
instance.